+++ packages/sharec/test/jest.matchers.js
@@ -2,6 +2,12 @@
 const diff = require('jest-diff')
 
 expect.extend({
+  /**
+   * This matcher requres for windows testsing and "lines" based configuration
+   * like YAML files or .gitignore
+   * There are a lot of different ways to do it correct, but in current moment -
+   * jest-matcher is the best and simple solution
+   */
   toWraplessEqual(rawReceived, rawExpected) {
     const options = {
       comment: 'Strings equality with ignoring of \\n and \\n\\r',